[C#] Recovery fixes (#421)
* [C#] Fix recovery of hash tables, speed up snapshot recovery.

Use snapshot as default checkpoint for benchmarks.

* avoid unnecessary hlog read.

* ensure tail page allocated after recovery

+++ cs/src/core/Allocator/BlittableAllocator.cs
@@ -130,7 +130,7 @@ namespace FASTER.core
 
             handles[index] = GCHandle.Alloc(tmp, GCHandleType.Pinned);
             long p = (long)handles[index].AddrOfPinnedObject();
-            pointers[index] = (p + (sectorSize - 1)) & ~(sectorSize - 1);
+            pointers[index] = (p + (sectorSize - 1)) & ~((long)sectorSize - 1);
             values[index] = tmp;
         }
